Create separate network_blas_cl_common.h
