+++ lc0/src/neural/network_blas_cl_common.cc
@@ -17,88 +17,17 @@
     along with Leela Zero.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-/*
-get_input_channels() -> kInputPlanes, Network::initialize::channels -> ConvBlock.biases.size()
-conv_weights -> ConvBlock.weights, conv_biases -> ConvBlock.biases
-batchnorm_means -> ConvBlock.bn_means, batchnorm_stddivs -> ConvBlock.bn_stddivs
-conv_pol_w -> Weights.policy.weights, conv_pol_b -> Weights.policy.biases
-bn_pol_w1 -> Weights.policy.bn_means, bn_pol_w2 -> Weights.policy.bn_stddivs // ??? not very useful oldnames
-NUM_POLICY_INPUT_PLANES -> Weights.policy.bn_means.size()
-ip_pol_w -> Weights.ip_pol_w, ip_pol_b -> Weights.ip_pol_b
-get_num_output_policy() -> Weights.ip_pol_b.size()
-conv_val_w -> Weights.value.weights, conv_val_b -> Weights.value.biases
-bn_val_w1 -> Weights.value.bn_means, bn_val_w2 -> Weights.value.bn_stddivs
-NUM_VALUE_INPUT_PLANES -> Weights.value.bn_means.size()
-{ip1_val_w,ip1_val_b,ip2_val_w,ip2_val_b} -> Weights.{}
-NUM_VALUE_CHANNELS -> Weights.ip1_val_b.size()
-
-*/
-
-#pragma once
-
-#include "factory.h" // network.h, optionsdict.h
+#include "network_blas_cl_common.h"
 
 namespace lczero {
 
-class BlasCLNetworkComputation : public NetworkComputation {
-
-  BlasCLNetworkComputation(const BlasCLNetwork* network) : network(network) {}
-
-  void AddInput(InputPlanes&& input) override {
-    inputs.push_back(input);
-  }
-
-  void GetBatchSize() override {
-    return inputs.size();
-  }
-
-  void ComputeBlocking() override {
-    output_values = std::make_unique<float[]>(inputs.size());
-    output_policies = std::make_unique<float[][]>(inputs.size());
-    for (size_t i = 0; i < inputs.size(); i++)
-      output_values[i], output_policies[i] = network->evaluate(inputs[i]);
-  }
-  
-  int GetQVal(int sample) override {
-    return output_values[sample];
-  }
-  
-  float GetPVal(int sample, int move_id) override {
-    return output_policies[sample][move_id];
-  }
-
- private:
-  std::vector<InputPlanes> inputs;
-  float[]   output_values;
-  float[][] output_policies;
-  BlasCLNetwork* network;
+void BlasCLNetworkComputation::ComputeBlocking() override {
+  output_values = std::make_unique<float[]>(inputs.size());
+  output_policies = std::make_unique<float[][]>(inputs.size());
+  for (size_t i = 0; i < inputs.size(); i++)
+    output_values[i], output_policies[i] = network->evaluate(inputs[i]);
 }
-
-class BlasCLNetwork : public Network {
- public:
-  BlasCLNetwork(const Weights& weights, const OptionsDict& options)
-    : weights(weights), options(options) {
-    initialize();
-  }
-    
-  std::unique_ptr<NetworkComputation> NewComputation() override {
-    return std::make_unique<BlasCLNetworkComputation>(this);
-  }
   
-  virtual std::pair<float value, float[] policy> evaluate(InputPlanes&& input);
-  // output array should be std::make_unique  
-  
- protected:
-  virtual void forward(...);
-  void softmax();
- 
-  void initialize(void);
-  void initOneBlock(Weights::ConvBlock& block, bool inputlayer=false);
-  Weights::Vec winograd_transform_f(const Weights::Vec& f, const int outputs, const int channels);
-  
-  Weights weights; // optimal memory use? is one reference shared among multiple backends?
-  OptionsDict& options;
-}
 
 void BlasCLNetwork::initialize(void) {
   // this matches old Network::initialize, in Network.cpp:375-416   
@@ -153,4 +82,8 @@ std::pair<float value, float[] policy> evaluate(InputPlanes&& input) {
   return {value, policy};
 }
 
-void BlasCLNetwork::softmax();
+} // namespace lczero
+
+void BlasCLNetwork::softmax() {
+
+}
